+++ DLL/LouDLLs/User32/Interface/WindowCreation/Styles.c
@@ -3,15 +3,15 @@
 __declspec(dllimport)
 void LouDrsdDrawLine32(
     PDRSD_CLIP Clip, 
-    UINT32 X1, UINT32 Y1,
-    UINT32 X2, UINT32 Y2, 
+    INT64 X1, INT64 Y1,
+    INT64 X2, INT64 Y2, 
     UINT32 Color
 );
 
 __declspec(dllimport)
 void LouDrsdPutPixel32(
     PDRSD_CLIP Clip,
-    UINT32 X, UINT32 Y,
+    INT64 X, INT64 Y,
     UINT32 Color
 );
